Move agent icons to @upstream/common package
- Created packages/common/src/agent-icons.tsx with all icon components
- Added React as peer dependency to common package
- Added JSX support to common package tsconfig
- Updated web and simple-chat to re-export icons from common
- Eliminates code duplication between packages
