+++ packages/common/package.json
@@ -18,8 +18,17 @@
     "clsx": "^2.1.0",
     "tailwind-merge": "^2.2.0"
   },
+  "peerDependencies": {
+    "react": "^18.0.0 || ^19.0.0"
+  },
+  "peerDependenciesMeta": {
+    "react": {
+      "optional": true
+    }
+  },
   "devDependencies": {
     "@types/node": "^20.10.0",
+    "@types/react": "^18.0.0",
     "typescript": "^5.3.0"
   }
 }
